+++ R/class_layers.R
@@ -124,7 +124,6 @@ LayersR6 <- R6::R6Class(
     .get_py_object = function() {
       private$.layers
     }
-
   ),
   active = list(
     #' @field parent Reference to parent AnnData view
@@ -143,6 +142,7 @@ LayersR6 <- R6::R6Class(
 #'
 #' @rdname LayersHelpers
 #' @export
+#' @method names LayersR6
 #'
 #' @examples
 #' \dontrun{
@@ -167,30 +167,35 @@ names.LayersR6 <- function(x) {
 }
 
 #' @rdname LayersHelpers
+#' @method length LayersR6
 #' @export
 length.LayersR6 <- function(x) {
   x$length()
 }
 
 #' @rdname LayersHelpers
+#' @method r_to_py LayersR6
 #' @export
 r_to_py.LayersR6 <- function(x, convert = FALSE) {
   x$.get_py_object()
 }
 
 #' @rdname LayersHelpers
+#' @method py_to_r anndata._core.aligned_mapping.LayersBase
 #' @export
 py_to_r.anndata._core.aligned_mapping.LayersBase <- function(x) {
   LayersR6$new(x)
 }
 
 #' @rdname LayersHelpers
+#' @method [ LayersR6
 #' @export
 `[.LayersR6` <- function(x, name) {
   x$get(name)
 }
 
 #' @rdname LayersHelpers
+#' @method [<- LayersR6
 #' @export
 `[<-.LayersR6` <- function(x, name, value) {
   x$set(name, value)
@@ -198,16 +203,18 @@ py_to_r.anndata._core.aligned_mapping.LayersBase <- function(x) {
 }
 
 #' @rdname LayersHelpers
+#' @method [[ LayersR6
 #' @export
 `[[.LayersR6` <- `[.LayersR6`
 
 #' @rdname LayersHelpers
+#' @method [[<- LayersR6
 #' @export
 `[[<-.LayersR6` <- `[<-.LayersR6`
 
-
 #' @rdname all.equal
-#' @export
+#' @method all.equal LayersR6
+#' @exportS3Method all.equal LayersR6
 all.equal.LayersR6 <- function(target, current, ...) {
   a <- target
   b <- current